refactor: make import of components consistent
