+++ client/components/index.js
@@ -22,3 +22,4 @@ export {default as InboxComponent} from './chat/InboxComponent'
 export {default as InboxOnly} from './chat/InboxOnly'
 export {default as ItemForm} from './ItemForm'
 export {default as EditItemForm} from './EditItemForm'
+export {default as MobileFooter} from './MobileFooter'
